requirements.txt
torch
torchvision
transformers
avalanche-lib
matplotlib
numpy
pandas
ftfy
tqdm
regex
scikit-learn
gdown